Negative Distance Scoring Fix
Used to normalize the value of maxDistance-distance, but between the minDistance and maxDistance. The difference value was kept in because we used to always just normalize from 0 to whatever. But now that we track miniumum values, we can be more precise. The distance itself is directly normalized FIRST, then taken away from 1, so same idea of shortest distance with largest score. Minimum distance score is to 0.05 (1-0.95). No more weird negative scores.