Add nearby sorting function
Similar to our previous sort locations function but now can specify radius within which we want places. Currently defaulted to (0.1, unclear miles or km). The benefit to this is that we can use it for our check-in feature, which will bring up all the nearest places within which you can possible check-in from. Changes made with accompanying testing code in vibemap-app/actions.js